+++ bytes/doc.go
@@ -1,9 +1,0 @@
-// Copyright 2018 Miguel Angel Rivera Notararigo. All rights reserved.
-// This source code was released under the MIT license.
-
-/*
-Package bytes implements functions for manipulating byte slices.
-*/
-package bytes
-
-// API Status: stable